requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
bcrypt==4.0.1
sqlite3
datetime
fpdf2==2.7.6
matplotlib==3.8.0
Pillow==10.0.1
reportlab==4.0.5
gunicorn==21.2.0
python-dotenv==1.0.0 